requirements.txt
pandas
gensim
streamlit
pyyaml
